rust: use SMART_KEYMAP_CUSTOM_KEYMAP for custom keymap
